+++ shared/locales/en-US.js
@@ -33,7 +33,7 @@ module.exports = {
     lost_internet_connection: "\nLost internet connection:\n%(output)s",
     connect_docker_unavailable: "Could not initialize balancer because docker was not available",
     agent_not_running: "azk agent is required but is not running (try `azk agent status`)",
-    agent_start: "azk agent start error: %(error)s",
+    agent_start: "azk agent start error: %(err_message)s",
     agent_stop:  "azk agent stop error (try `azk agent status`)",
     not_been_implemented: "This feature: `%(feature)s` has not been implemented yet",
     system_not_found: "System `%(system)s` not found in `%(manifest)s`",
@@ -67,7 +67,8 @@ module.exports = {
 
     docker_build_error: {
       command_error: "  Error in building `%(dockerfile)s`:\n%(output)s\n",
-      server_error: "Internal error in build `%(dockerfile)s`: %(error)",
+      server_error: "Internal error in build `%(dockerfile)s`: %(error)s",
+      unknow_instrction_error: "Unknown instruction in build `%(dockerfile)s`: %(instruction)s",
       not_found   : "Can't find `%(from)s` image to build `%(dockerfile)s`",
       can_find_dockerfile: "Can't find `%(dockerfile)s` file",
       can_find_add_file_in_dockerfile: "Can't find `%(source)s` file to ADD in `%(dockerfile)s`",
@@ -412,17 +413,18 @@ module.exports = {
       }
     },
     scale: {
-      instances   : "from " + "%(from)d".red + " to " + "%(to)d".green + " instances",
+      instances   : "from " + "%(from)d".red             + " to " + "%(to)d".green + " instances",
       description : "Scales (up or down) an instance of the system(s)",
-      wait_port   : "◴".magenta + " waiting for `" + "%(system)s".blue  + "` system to start, trying connection to port %(name)s/%(protocol)s...",
-      check_image : "✓".cyan    + " checking `"      + "%(image)s".yellow + "` image...",
-      pull_image  : "⇲".blue    + " downloading `"   + "%(image)s".yellow + "` image...",
-      build_image : "⇲".blue    + " building `"      + "%(image)s".yellow + "` image...",
-      provision   : "↻".yellow  + " provisioning `"  + "%(system)s".blue  + "` system...",
-      starting    : "↑".green   + " starting `"      + "%(system)s".blue  + "` system, " + "%(to)d".green + " new instances...",
-      stopping    : "↓".red     + " stopping `"      + "%(system)s".blue  + "` system, " + "%(from)d".red + " instances...",
-      scaling_up  : "↑".green   + " scaling `"       + "%(system)s".blue  + "` system %(instances)s...",
-      scaling_down: "↓".red     + " scaling `"       + "%(system)s".blue  + "` system %(instances)s...",
+      sync        : "⎘".yellow  + " syncing files for `" + "%(system)s".blue     + "` system...",
+      wait_port   : "◴".magenta + " waiting for `"       + "%(system)s".blue     + "` system to start, trying connection to port %(name)s/%(protocol)s...",
+      check_image : "✓".cyan    + " checking `"          + "%(image)s".yellow    + "` image...",
+      pull_image  : "⇲".blue    + " downloading `"       + "%(image)s".yellow    + "` image...",
+      build_image : "⇲".blue    + " building `"          + "%(image)s".yellow    + "` image...",
+      provision   : "↻".yellow  + " provisioning `"      + "%(system)s".blue     + "` system...",
+      starting    : "↑".green   + " starting `"          + "%(system)s".blue     + "` system, " + "%(to)d".green + " new instances...",
+      stopping    : "↓".red     + " stopping `"          + "%(system)s".blue     + "` system, " + "%(from)d".red + " instances...",
+      scaling_up  : "↑".green   + " scaling `"           + "%(system)s".blue     + "` system %(instances)s...",
+      scaling_down: "↓".red     + " scaling `"           + "%(system)s".blue     + "` system %(instances)s...",
       options: {
         remove: "Removes the instances before stopping",
         verbose: verbose,
@@ -552,7 +554,7 @@ module.exports = {
       },
     },
   },
-  
+
   tracking: {
     timeout: "Analytics tracker timed out."
   }